Extract path and rename UI to EditorPanelHeaderPath
- Offloads path rendering, renaming, and context-menu behaviors from
  EditorPanelHeader into a new dedicated component.
- Improves code maintainability and readability by decoupling distinct
  responsibilities within the editor panel headers.
